% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/50_dfbeta-numeric.R
\name{dfbeta_numeric}
\alias{dfbeta_numeric}
\title{Set DFBETA for numeric inputs}
\usage{
dfbeta_numeric(y, X, set, col_X = 1)
}
\arguments{
\item{y}{Numeric vector of response values}

\item{X}{Matrix or data frame of predictor variables}

\item{set}{Vector of indices for the set whose influence is being assessed}

\item{col_X}{Column index of X to compute DFBETA for (default: 1)}
}
\value{
A numeric value representing the DFBETA for observations in S
}
\description{
Calculates the DFBETA for a set of observations (S)
}
