% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_estimate_dfb_evd.R
\name{estimate_dfb_evd}
\alias{estimate_dfb_evd}
\title{Estimate EVD of Set Influence}
\usage{
estimate_dfb_evd(y, x, Z, set, block_count = 20, verbose = TRUE)
}
\arguments{
\item{y}{Response variable vector}

\item{x}{Primary predictor variable of interest}

\item{Z}{Matrix of other predictor variables to be marginalized out}

\item{set}{Vector of indices for observations whose influence is being assessed}

\item{block_count}{Number of blocks to divide the data into}

\item{verbose}{Logical indicating whether to print results}
}
\value{
List containing the estimated EVD parameters, set DFBETA, and block maxima
}
\description{
Estimate the extreme value distribution (EVD) parameters of set influence.
Uses the Frisch-Waugh-Lovell theorem to marginalize out other predictors
and fits a generalized extreme value distribution to block maxima.
}
