% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/20_rdfbeta.R
\name{rdfbeta}
\alias{rdfbeta}
\title{Random Maximum Influence Draw}
\usage{
rdfbeta(
  n_draw = 10L,
  n = 100L,
  n_set = 1L,
  x_dist = function(n) rnorm(n),
  r_dist = function(n) rnorm(n)
)
}
\arguments{
\item{n_draw}{Number of draws}

\item{n}{Sample size (default: 100)}

\item{n_set}{Number of most influential observations to select (default: 1)}

\item{x_dist}{Function for generating X values (default: standard normal)}

\item{r_dist}{Function for generating residuals (default: standard normal)}
}
\value{
Numeric value of the set influence in a random sample
}
\description{
Generates randomDFBETA values by simulating data.
}
