% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/21_rmaxdfbeta.R
\name{rmaxdfbeta}
\alias{rmaxdfbeta}
\title{Random Maximum Influence Draw}
\usage{
rmaxdfbeta(
  n = 100L,
  n_set = 1L,
  x_dist = function(n) rnorm(n),
  r_dist = function(n) rnorm(n)
)
}
\arguments{
\item{n}{Sample size (default: 100)}

\item{n_set}{Number of most influential observations to select (default: 1)}

\item{x_dist}{Function for generating X values (default: standard normal)}

\item{r_dist}{Function for generating residuals (default: standard normal)}
}
\value{
Numeric value of the maximum set influence in a random sample
}
\description{
Generates random maximum DFBETA values by simulating data and iteratively
selecting the most influential observations in a greedy manner for
theoretical distribution analysis of maximum influence measures.
}
